Improve 1D integration test w.r.t. to precision
The GSL integration wrapper was not really properly used in the unit
tests. In particular, simple functions were integrated for which the GSL
error was estimated orders of magnitute smaller than the requested
accuracy. Therefore, comparing the exact result and the calculated one
using the GSL estimated error as absolute threshold was too strict and
failing on some machines. Now I expect the result to be exact at most 3
ULP and I compare in an absolute sense at 2 sigma. This might in
principle still fail, but it should be quite conservative already.