Rename HyperSurfaceCrossing to Fluidization and add new process type that doesnt remove particles
Add "fluidized" to ParticleData.

It becomes true once the particle fluidizes, and there is no mechanism to make it false. Moreover, this trait is inherited by all children of this particle (whether via decay, resonance, or string process), meaning that they cannot themselves become fluid. That is, only the original particle will be written to the IC output.

The idea behind this is to have a fluidization without removing the particles from the simulation, since they should contribute to the energy density in the case of dynamic fluidization.

For now this is hardcoded but will become an option.

Make fluidization not count as a collision.

Otherwise products of a collision can fluidize and re-interact, leading to an infinite binary interaction loop