Set `time_last_coll` to formation time for created particles
For every newly created particle the time of last collision
`time_last_coll` was set to zero at first. Now the `formation_time`, the
`begin_formation_time`, and the `time_last_coll` are initialized as
`NAN` and when the `time_last_coll` is `NAN` when the formation time is
set then it is set to the formation time of the newly created particle.