Fix logger message for start time in fixed output time steps
If the start time was given in fixed output time steps, there was a
warning about this time step being removed from the output times.
Although, the removal is the right action to take, the message was
confusing, hence this message for the start time is now changed and only
visible in debug mode of the logger. Additionally, I added a test
about setting the output interval to 0.