Fix documentation for AQM suppression keys of heavy flavor
The documentation was incorrect regarding the AQM suppression keys for
bottom and charm hadrons. The documentation put them in the
`Collision_Term` section, leaving out that they should be entered within
the subsection `Heavy_Flavor`. Furthermore, the default value of
`AQM_Bottom_Suppression` that was stated in the documentation was wrong.