Fix `frac_bottom` function returning non-zero for s quarks
The first if-statement in the function `frac_bottom` was true if the
hadron contained two strange quarks, not two bottom quarks.
Additionally, re-structured the if-else statements in all frac functions
to consider `is_meson` when checking for quarkonium instead of
`is_hadron`. This is less error-prone since baryons nevertheless do not
contain a quark and anti-quark pair.