Update Ubuntu OS setup for actions on GitHub servers
As ubuntu-20.04 is not supported any more by GitHub in actions, we
needed to rethink our setup. Now older GNU and LLVM compilers are used
with ubuntu-22.04 and the default ones are used with ubuntu-24.04. Since
it is not possible to install the SMASH oldest supported compiler as
package on more recent ubuntu distributions, we decided not to do so in
the public GitHub actions.