Store the Key::validator_ as non-optional std::function
Instead of adding an std::optional layer, it is simply possible to
always store a validator functor in the Key class which is either what
the user specifies or a bool function that always returns true. This
simplifies the code a bit as the validator can always be used without
testing the std::optional state.