Limit overloads of operator<< for printable types
In the grid test file there are overloads of the operator<< added to the
std namespace to print an std::pair, an std::set, an std::unordered_set
and an std::vector<std::pair<>>. These were just templated on the
contained types and there was no constraint on the type being streamable
to the std::ostream object. Being those overload added to std namespace,
which is strictly speaking UB, this was potentially opening up for a
call to those << operators with containers containing types which could
not be printed. Actually, this precisely happened compiling Key types
with e.g. std::vector<enum> as key type. Key::as_yaml method streams
the key value via << to an std::stringstream only if the key type has
such an operator overload implemented. This method is now used in
Configuration::take and Configuration::read to print the error message
if the key value is not valid. In general std::vector has no <<
overload, but for the grid tests this was added into std namespace
without taking care to limit it to fully streamable types. Hence
compilation of the grid test file broke. This commit fixes the problem
using std::enable_if as additional template parameter in the overloads
of operator<< to restrict the template to fully streamable objects.