Enforce noexcept specification of Key::validate method
Throwing validators are now accepted from the Key class as in C++17 it
would be too much hassle to avoid it. In C++23 there are
std::function-like wrappers that are specialised for noexcept functions
and this can be then changed in the future. At the moment the validate
method is considering the passed value invalid and returns false if the
validator throws an exception. This behaviour is considered a misuse of
the class and a non-error is given to the user.