Fix subtle Key bug due to static initialisation order fiasco
This bug was present basically since the introduction of the validator,
but was now fortunately caught by the sanitiser crashing in a totally
unrelated test on a given machine only. The problem is that, although in
principle no InputKey was passing a validator to the Key constructor and
the default one should have always been used, the default validator was
returned by a static function returning a function-local static and,
basically, the functor could have left the function zero initialised,
due to undefined order of static initialisation. This empty validator,
in turn, was causing the Key constructor to throw, hence unwinding
during static initialization which made the sanitiser crash (exceptions
thrown during static initialisation cannot be caught as this happens
before any main code is run).

Some further detail to clarify further why the default validator could
be returned empty. Remember that C++ splits initialization rules of
non-local static objects into two phases: a first phase that involves
zero-initialization of storage before any code runs and then a second
one which executes the constructor calls (which may run code). There is
no defined ordering of the second phase between different translation
units (TUs), or between inline static objects in different headers that
may be included by multiple TUs. Now, in our code, C++ guarantees that
the function-local variable "always_true_functor" is initialized on the
first call to the function, in a thread-safe way, the first time control
passes through its declaration. However, at the moment that some other
static Key is being dynamically initialized, it’s possible that:
 1. get_default_validator_() is being called as part of another static
    initialization, and inside that, the runtime is in the middle of
    initializing function-local statics in the same or other TUs;
 2. due to interleaving, the internal static always_true_functor's
    memory exists (zero-initialized), but its constructor (the lambda’s
    construction) hasn’t yet run when the current thread tries to copy
    it.
In that state, copying from always_true_functor results in a
default-constructed std::function (empty callable), because the internal
lambda object hasn’t yet been fully constructed.

All this story made me realise that the Key constructor could also get
an explicit empty std::function (because the caller got it wrong or who
knows why) and hence it is just fine and important to ensure that the
validator_ member is not empty.

In this commit I also refactored the function returning the default
validator outside the class into the detail namespace as a function
template. This slightly reduces complexity in terms of static
initialisation and disentangle a bit the code. The inner function-local
static functor remain, though. I also decided to return the functor by
const reference, not because it really changes a lot (returning by value
makes the return statement create an std::function that then can be
moved at the caller site, while returning a const& avoid the creation of
the std::function but forces the caller to copy the returned referred
std::function), but because it makes the code a bit easier to reason
about: For each Key type there is a single always_true_functor on static
storage.