Implement alternative approach for root finding
In the momentum dependence approach four symmetric intervals around an
educated initial guess were attempted to find the root. This was failing
in some setup. Now a new approach is proposed. The algorithm builds
larger and larger not necessarily symmetric intervals around the
educated initial guesses, if the function of which a root should be
found does not have opposite signs at the boundaries. This scan is done
inside an interval of a fixed maximum width around the educated guess
and it is repeated three times with always finer resolution if a
previous scan failed.