Reintroduce old implementation and use new one as plan B
Since I do not have too much control over the potentials implementation
and playing with the scanning resolutions and starting interval I was
getting warnings that I did not expect (probably from the function
evaluation), I decided to reintroduce the old implementation and add the
new one as plan B. In the future this can be evaluated and changed
again. The other advantage is that performance is not affected, in the
sense that the older implementation will still be the one executed for
all runs that did not fail. Now runs that previously failed might manage
to continue.